Consultas preparadas en PDO

+++ ra5/bbdd/mysqli/07baja_cliente.php
@@ -9,7 +9,8 @@ echo "<header>Baja de cliente</header>";
 
 // Verificar el token
 if( isset($_COOKIE['jwt']) ) {
-    $payload = $_COOKIE['jwt'];
+    $jwt = $_COOKIE['jwt'];
+    $payload = verificar_token($jwt);
 
     if( !$payload) {
         header("Location: 04autenticacion.php");
@@ -25,24 +26,28 @@ if( $_SERVER['REQUEST_METHOD'] == "GET") {
 <form method="POST" action="<?=$_SERVER['PHP_SELF']?>">
     <input type="submit" name="operacion" id="operacion" value="Darme de baja">
 </form>
+<p><a href="05pagina_inicio.php">No quiero de darme de baja. Me voy a la zona autenticada</a></p>
 <?php
 }
 elseif( $_SERVER['REQUEST_METHOD'] == "POST") {
     $operacion = filter_input(INPUT_POST, "operacion", FILTER_SANITIZE_SPECIAL_CHARS);
 
-    try {
-        $cbd = new mysqli($servidor_bd, $usuario, $clave, $esquema, $puerto);
+    if( $operacion == "Darme de baja" ) {
+        try {
+            $cbd = new mysqli($servidor_bd, $usuario, $clave, $esquema, $puerto);
 
-        $sql = "DELETE FROM cliente WHERE nif = ?";
-        $stmt = $cbd->prepare($sql);
-        $stmt->bind_param("s", $payload['nif']);
-        if( $stmt->execute() && $stmt->affected_rows == 1 ) {
-            echo "<h3>Cliente {$payload['nombre']} ha sido dado de baja.</h3>";
-            cerrar_sesion();
+            $sql = "DELETE FROM cliente WHERE nif = ?";
+            $stmt = $cbd->prepare($sql);
+            $stmt->bind_param("s", $payload['nif']);
+            if( $stmt->execute() && $stmt->affected_rows == 1 ) {
+                echo "<h3>Cliente {$payload['nombre']} ha sido dado de baja.</h3>";
+                echo "<p><a href='04registro_cliente.php'>Quiero registrarme otra vez</a></p>";
+                cerrar_sesion();
+            }
         }
-    }
-    catch( mysqli_sql_exception $mse) {
-        mostrar_error($mse);
+        catch( mysqli_sql_exception $mse) {
+            mostrar_error($mse);
+        }
     }
 }
 fin_html();
